requirements.txt
# UI
gradio==4.37.2

# LangChain core & integrations
langchain==0.3.26
langchain-google-genai==2.1.10    # LangChain integration for Google GenAI / Gemini
google-genai>=1.7.0               # Official Google GenAI SDK (Gemini)

# Pinecone (use the 'pinecone' package; do NOT keep 'pinecone-client' alongside it)
pinecone>=7.0.0

# LangChain - Pinecone integration (as you had)
langchain-pinecone==0.2.8

# Optional LangChain extras (keep if you use OpenAI bindings / community helpers)
langchain-openai==0.3.24
langchain-community==0.3.26

# Embeddings / Transformers / Torch
sentence-transformers==2.2.2
transformers==4.34.0
torch>=1.13.1

# Utilities
python-dotenv==1.0.1
pypdf==5.6.1
